xiaohongshu-generator/requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
jinja2>=3.0.0
playwright>=1.40.0
flask>=2.0.0
